requirements.txt
# Core Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
redis>=4.6.0

# Streaming & Message Queue
kafka-python>=2.0.2

# Web Scraping & HTTP
google-play-scraper>=1.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Natural Language Processing & Machine Learning
transformers>=4.33.0
torch>=2.0.0
tokenizers>=0.13.0
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
textblob>=0.17.0

# Deep Learning (Keras)
keras>=2.13.0
tensorflow>=2.13.0

# Dashboard & Visualization
streamlit>=1.26.0
plotly>=5.15.0
altair>=5.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Workflow Orchestration (Optional for Phase 1)
# apache-airflow>=2.7.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
click>=8.1.0
python-dateutil>=2.8.2
pytz>=2023.3

# Async Support
asyncio>=3.4.3
aiofiles>=23.1.0

# Configuration & Logging
loguru>=0.7.0
structlog>=23.1.0

# Data Validation
pydantic>=2.1.0
marshmallow>=3.20.0

# Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.60.0